+++ db/migrate/20190309155029_create_puntodeentregas.rb
@@ -2,7 +2,7 @@ class CreatePuntodeentregas < ActiveRecord::Migration[5.0]
   def change
     create_table :puntodeentregas do |t|
       t.string :nombre
-      t.numeric :habitantesBeneficiados
+      t.integer :habitantesBeneficiados
       t.text :ubicacion
       t.datetime :created_at
       t.datetime :updated_at
